добавлен тест для reader

+++ demod.h
@@ -23,9 +23,9 @@ private:
 
 template < typename Type >
 std::vector< Type > Demodulation::demodAM( SignalFile< Type > const& sigFile ) {
-    uint32_t size = sigFile.size();
+    uint64_t size = sigFile.size();
     std::vector< Type > data( size );
-    for( uint32_t i = 0; i < size; i++ ) {
+    for( uint64_t i = 0; i < size; i++ ) {
         data[ i ] = sigFile.data( i ).abs() - 1;
     }
     return std::move( data );
@@ -33,15 +33,15 @@ std::vector< Type > Demodulation::demodAM( SignalFile< Type > const& sigFile ) {
 
 template < typename Type >
 std::vector< Type > Demodulation::demodFM( SignalFile< Type > const& sigFile ) {
-    uint32_t size = sigFile.size();
-    int64_t fCen = sigFile.freqCenter();
-    int64_t fSam = sigFile.freqSample();
+    uint64_t size = sigFile.size();
+    uint64_t fCen = sigFile.freqCenter();
+    uint32_t fSam = sigFile.freqSample();
     std::vector< Type > data( size );
-    for( uint32_t i = 0; i < size; i++ ) {
+    for( uint64_t i = 0; i < size; i++ ) {
         data[ i ] = sigFile.data( i ).arg();
     }
 
-    for( uint32_t i = 0; i < size - 1; i++ ) {
+    for( uint64_t i = 0; i < size - 1; i++ ) {
         Type dif = ( data[ i + 1 ] - data[ i ] ) * 0.5 * M_1_PI;
         if( dif > 0.5 ) dif--;
         if( dif < -0.5 ) dif++;
