+++ writer.h
@@ -18,7 +18,7 @@ public:
      * @param ext используемый тип данных [s - short; f - float]
      */
     template < typename Type >
-    void writeFile( const std::string& path, std::vector< Type > const& data, int64_t fd );
+    void writeFile( const std::string& path, std::vector< Type > const& data, uint32_t fd );
 
 
 private:
@@ -27,14 +27,13 @@ private:
 
 
 template < typename Type >
-void Writer::writeFile( const std::string& path, std::vector< Type > const& data, int64_t fd ) {
+void Writer::writeFile( const std::string& path, std::vector< Type > const& data, uint32_t fd ) {
     std::string ext = typeid( data[ 0 ] ).name();
-    uint32_t size = data.size();
+    uint64_t size = data.size();
     std::string path_ = ( path + std::to_string( fd ) + ".iq" + ext );
     std::ofstream file( path_ );
-    for( uint32_t i = 0; i < size; i++ ) {
+    for( uint64_t i = 0; i < size; i++ ) {
         file.write( ( char* )&( data[ i ] ), sizeof( Type ) );
-
     }
 
     file.close();
